7_my-gpt/requirements copy.txt
langchain
chromadb
urllib3
PyMuPDF
python-dotenv
unstructured
openai
tiktoken
streamlit
